Add initial Django setup
Break the usual `settings.py` module into multiple modules in a `settings`
package in order to ease configuration in various contexts. Note that this
means the `DJANGO_SETTINGS_MODULE` environment variable will need to be set
in order for the app to work correctly.

Further, store sensitive values in a file named `BASE_DIR / ".env"`. This
file will need to be present and contain the following variables for the app
to work correctly:

* DJANGO_SECRET_KEY

The production version of `.env` must contain these variables in addition to
those listed above:

* DJANGO_DB_NAME
* DJANGO_DB_USER
* DJANGO_DB_PASSWORD