Add initial Braille app
A simple app that demonstrates the Django app installation process, including
using multiple levels of template inheritance and static files. The single-page
app allows users to use a QWERT keyboard as if it were a Perkins Brailler.

Update .gitignore to exclude `static/` in the root directory, as this is where
copies of all applications' static files are made by `manage.py collectstatic`
in production. There is no need to store these files twice.