Condense the Braille index page
Replace the previously verbose how-to notes with an accessibly
collapsable region that will initially hide this information for a
smoother user experience.

Modify the base and meta templates to include Bootstrap and its plugins
correctly. These are needed for the collapsing button on this page and
will doubtless be used widely throughout the project later.

Add unit tests for the Braille app.