Replace header template with a navbar
For easy navigation and organization of site demos-to-be, replace a simple
heading with a navbar which includes currently the following:

* A brand link that currently links to the site's homepage. Perhaps a full
  contact-us page might be better.
* A direct link to the site's homepage.
* A dropdown menu whose purpose is to display site-based demos. The only one
  of these currently is the Braille demo.
* A link to sign in / sign up.

Also add a proper top-level heading to the start of the `main` region of
the base template. This is essentially what the header template included
before, but moved down slightly below the navbar.